refactor(folio): move the layout pipeline compute into the controller
Extract runLayoutPipeline's body (and its nested helpers) verbatim into
core/controller/layoutPipeline.ts as a framework-agnostic function taking an
explicit deps object and returning a LayoutOutcome. The React adapter keeps a
thin wrapper that builds deps from current props/refs and applies the outcome
(setBlocks/setMeasures/setLayout/painter.setBlockLookup) in the original order.
Render/font helpers are injected as callbacks; engine stages are imported
directly. No behavior change.
